pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Yurb", email = "yavurb@gmail.com"},
]
dependencies = [
    "fastapi>=0.103.2",
    "uvicorn[standard]>=0.23.2",
    "pydantic-settings>=2.0.3",
    "python-dotenv>=1.0.0",
    "psycopg2>=2.9.8",
    "psycopg2-binary>=2.9.8",
    "sqlalchemy>=2.0.21",
    "alembic>=1.12.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
_.env_file = ".env"
start = "sh ./start.sh"

# Alembic - migrations commands
db_revision = "alembic revision --autogenerate -m"
db_upgrade = "alembic upgrade head"
db_downgrade = "alembic downgrade -1"
